+++ pybamm/processed_variable.py
@@ -39,14 +39,14 @@ class ProcessedVariable(object):
             self.base_eval, self.known_evals[solution.t[0]] = base_variable.evaluate(
                 solution.t[0],
                 solution.y[:, 0],
-                u={name: inp[0] for name, inp in solution.inputs.items()},
+                inputs={name: inp[0] for name, inp in solution.inputs.items()},
                 known_evals=self.known_evals[solution.t[0]],
             )
         else:
             self.base_eval = base_variable.evaluate(
                 solution.t[0],
                 solution.y[:, 0],
-                u={name: inp[0] for name, inp in solution.inputs.items()},
+                inputs={name: inp[0] for name, inp in solution.inputs.items()},
             )
 
         # handle 2D (in space) finite element variables differently
@@ -107,10 +107,10 @@ class ProcessedVariable(object):
             inputs = {name: inp[idx] for name, inp in self.inputs.items()}
             if self.known_evals:
                 entries[idx], self.known_evals[t] = self.base_variable.evaluate(
-                    t, u, u=inputs, known_evals=self.known_evals[t]
+                    t, u, inputs=inputs, known_evals=self.known_evals[t]
                 )
             else:
-                entries[idx] = self.base_variable.evaluate(t, u, u=inputs)
+                entries[idx] = self.base_variable.evaluate(t, u, inputs=inputs)
 
         # No discretisation provided, or variable has no domain (function of t only)
         self._interpolation_function = interp.interp1d(
@@ -131,12 +131,12 @@ class ProcessedVariable(object):
             inputs = {name: inp[idx] for name, inp in self.inputs.items()}
             if self.known_evals:
                 eval_and_known_evals = self.base_variable.evaluate(
-                    t, u, u=inputs, known_evals=self.known_evals[t]
+                    t, u, inputs=inputs, known_evals=self.known_evals[t]
                 )
                 entries[:, idx] = eval_and_known_evals[0][:, 0]
                 self.known_evals[t] = eval_and_known_evals[1]
             else:
-                entries[:, idx] = self.base_variable.evaluate(t, u, u=inputs)[:, 0]
+                entries[:, idx] = self.base_variable.evaluate(t, u, inputs=inputs)[:, 0]
 
         # Process the discretisation to get x values
         nodes = self.mesh[0].nodes
@@ -236,7 +236,7 @@ class ProcessedVariable(object):
             inputs = {name: inp[idx] for name, inp in self.inputs.items()}
             if self.known_evals:
                 eval_and_known_evals = self.base_variable.evaluate(
-                    t, u, u=inputs, known_evals=self.known_evals[t]
+                    t, u, inputs=inputs, known_evals=self.known_evals[t]
                 )
                 entries[:, :, idx] = np.reshape(
                     eval_and_known_evals[0],
@@ -246,7 +246,7 @@ class ProcessedVariable(object):
                 self.known_evals[t] = eval_and_known_evals[1]
             else:
                 entries[:, :, idx] = np.reshape(
-                    self.base_variable.evaluate(t, u, u=inputs),
+                    self.base_variable.evaluate(t, u, inputs=inputs),
                     [first_dim_size, second_dim_size],
                     order="F",
                 )
@@ -275,7 +275,7 @@ class ProcessedVariable(object):
         inputs = {name: inp[0] for name, inp in self.inputs.items()}
 
         entries = np.reshape(
-            self.base_variable.evaluate(0, self.u_sol, u=inputs), [len_y, len_z]
+            self.base_variable.evaluate(0, self.u_sol, inputs=inputs), [len_y, len_z]
         )
 
         # assign attributes for reference
@@ -308,13 +308,13 @@ class ProcessedVariable(object):
 
             if self.known_evals:
                 eval_and_known_evals = self.base_variable.evaluate(
-                    t, u, u=inputs, known_evals=self.known_evals[t]
+                    t, u, inputs=inputs, known_evals=self.known_evals[t]
                 )
                 entries[:, :, idx] = np.reshape(eval_and_known_evals[0], [len_y, len_z])
                 self.known_evals[t] = eval_and_known_evals[1]
             else:
                 entries[:, :, idx] = np.reshape(
-                    self.base_variable.evaluate(t, u, u=inputs), [len_y, len_z]
+                    self.base_variable.evaluate(t, u, inputs=inputs), [len_y, len_z]
                 )
 
         # assign attributes for reference