+++ tests/unit/test_parameters/test_parameter_values.py
@@ -130,12 +130,14 @@ class TestParameterValues(unittest.TestCase):
         # if only C-rate and capacity provided, update current
         values = {"C-rate": "[input]", "Cell capacity [A.h]": 10}
         param = pybamm.ParameterValues(values)
-        self.assertEqual(param["Current function [A]"](2).evaluate(u={"C-rate": 1}), 10)
+        self.assertEqual(
+            param["Current function [A]"](2).evaluate(inputs={"C-rate": 1}), 10
+        )
         # if only current and capacity provided, update C-rate
         values = {"Current function [A]": "[input]", "Cell capacity [A.h]": 10}
         param = pybamm.ParameterValues(values)
         self.assertEqual(
-            param["C-rate"](5).evaluate(u={"Current function [A]": 5}), 0.5
+            param["C-rate"](5).evaluate(inputs={"Current function [A]": 5}), 0.5
         )
 
     def test_process_symbol(self):
@@ -283,7 +285,7 @@ class TestParameterValues(unittest.TestCase):
         a = pybamm.Parameter("a")
         processed_a = parameter_values.process_symbol(a)
         self.assertIsInstance(processed_a, pybamm.InputParameter)
-        self.assertEqual(processed_a.evaluate(u={"a": 5}), 5)
+        self.assertEqual(processed_a.evaluate(inputs={"a": 5}), 5)
 
         # process binary operation
         b = pybamm.Parameter("b")
@@ -292,7 +294,7 @@ class TestParameterValues(unittest.TestCase):
         self.assertIsInstance(processed_add, pybamm.Addition)
         self.assertIsInstance(processed_add.children[0], pybamm.InputParameter)
         self.assertIsInstance(processed_add.children[1], pybamm.Scalar)
-        self.assertEqual(processed_add.evaluate(u={"a": 4}), 7)
+        self.assertEqual(processed_add.evaluate(inputs={"a": 4}), 7)
 
     def test_process_function_parameter(self):
         parameter_values = pybamm.ParameterValues(
@@ -308,7 +310,7 @@ class TestParameterValues(unittest.TestCase):
         # process function
         func = pybamm.FunctionParameter("func", a)
         processed_func = parameter_values.process_symbol(func)
-        self.assertEqual(processed_func.evaluate(u={"a": 3}), 369)
+        self.assertEqual(processed_func.evaluate(inputs={"a": 3}), 369)
 
         # process constant function
         const = pybamm.FunctionParameter("const", a)
@@ -319,7 +321,7 @@ class TestParameterValues(unittest.TestCase):
         # process differentiated function parameter
         diff_func = func.diff(a)
         processed_diff_func = parameter_values.process_symbol(diff_func)
-        self.assertEqual(processed_diff_func.evaluate(u={"a": 3}), 123)
+        self.assertEqual(processed_diff_func.evaluate(inputs={"a": 3}), 123)
 
         # function parameter that returns a python float
         func = pybamm.FunctionParameter("float_func", a)
@@ -331,7 +333,7 @@ class TestParameterValues(unittest.TestCase):
         a = pybamm.Scalar(3)
         func = pybamm.FunctionParameter("func", a)
         processed_func = parameter_values.process_symbol(func)
-        self.assertEqual(processed_func.evaluate(u={"func": 13}), 13)
+        self.assertEqual(processed_func.evaluate(inputs={"func": 13}), 13)
 
     def test_process_inline_function_parameters(self):
         def D(c):
@@ -343,12 +345,12 @@ class TestParameterValues(unittest.TestCase):
         func = pybamm.FunctionParameter("Diffusivity", a)
 
         processed_func = parameter_values.process_symbol(func)
-        self.assertEqual(processed_func.evaluate(u={"a": 3}), 9)
+        self.assertEqual(processed_func.evaluate(inputs={"a": 3}), 9)
 
         # process differentiated function parameter
         diff_func = func.diff(a)
         processed_diff_func = parameter_values.process_symbol(diff_func)
-        self.assertEqual(processed_diff_func.evaluate(u={"a": 3}), 6)
+        self.assertEqual(processed_diff_func.evaluate(inputs={"a": 3}), 6)
 
     def test_multi_var_function_with_parameters(self):
         def D(a, b):
@@ -415,8 +417,8 @@ class TestParameterValues(unittest.TestCase):
         processed_func = parameter_values.process_symbol(func)
         processed_interp = parameter_values.process_symbol(interp)
         np.testing.assert_array_almost_equal(
-            processed_func.evaluate(u={"a": 0.6}),
-            processed_interp.evaluate(u={"a": 0.6}),
+            processed_func.evaluate(inputs={"a": 0.6}),
+            processed_interp.evaluate(inputs={"a": 0.6}),
             decimal=4,
         )
 
@@ -426,8 +428,8 @@ class TestParameterValues(unittest.TestCase):
         processed_diff_func = parameter_values.process_symbol(diff_func)
         processed_diff_interp = parameter_values.process_symbol(diff_interp)
         np.testing.assert_array_almost_equal(
-            processed_diff_func.evaluate(u={"a": 0.6}),
-            processed_diff_interp.evaluate(u={"a": 0.6}),
+            processed_diff_func.evaluate(inputs={"a": 0.6}),
+            processed_diff_interp.evaluate(inputs={"a": 0.6}),
             decimal=2,
         )