+++ tests/unit/test_solvers/test_casadi_algebraic_solver.py
@@ -36,10 +36,10 @@ class TestCasadiAlgebraicSolver(unittest.TestCase):
             y0 = np.array([2])
             t = casadi.MX.sym("t")
             y = casadi.MX.sym("y")
-            u = casadi.MX.sym("u")
-            casadi_algebraic = casadi.Function("alg", [t, y, u], [y ** 2 + 1])
+            p = casadi.MX.sym("p")
+            casadi_algebraic = casadi.Function("alg", [t, y, p], [y ** 2 + 1])
 
-            def algebraic_eval(self, t, y):
+            def algebraic_eval(self, t, y, inputs):
                 # algebraic equation has no real root
                 return y ** 2 + 1